Linked the Code reviewers to Gerrit
Gerrit is the single place where the list of people with +2 permissions
are up to date. Linking from the website to Gerrit will make sure that
we see the most up-to-date list of people with +2 permissions.
As an added bonus, you can quickly glean the email addresses to start
poking them offline for pending reviews.

Change-Id: Ic2afee754148c533437b579341f0e303d138ba09
Signed-off-by: Rahul Rajagopal <rahulr@hpe.com>