Add dependency required by the webui
The webui uses phantomJS for testing, and this one requires fontconfig to work.
If you are using a headless Linux installation, this may not be installed.
The native version of the library is causing a segfault when mixed with the host
libraries, so it's better to call out the dependency directly.

TG-860 #fixed

Change-Id: Ia8040ed933bf425a8a95a49d3161f892f39cb819
Signed-off-by: Diego Dompe <dompe@hpe.com>